OnlineBankingCore: create new base online banking view
