+++ online-banking-core/src/main/res/layout/online_banking_cz_spinner_layout.xml
